// packages/bridge/src/enums/field-attribute.ts
export const enum FieldAttributeFlags {
    FIELD_ACCESS_MASK = 0x0007,
    PRIVATE_SCOPE = 0x0000,
    PRIVATE = 0x0001,
    FAMILY_AND_ASSEMBLY = 0x0002,
    ASSEMBLY = 0x0003,
    FAMILY = 0x0004,
    FAMILY_OR_ASSEMBLY = 0x0005,
    PUBLIC = 0x0006,
    STATIC = 0x0010,
    INIT_ONLY = 0x0020,
    LITERAL = 0x0040,
    NOT_SERIALIZED = 0x0080,
    SPECIAL_NAME = 0x0200,
    PINVOKE_IMPL = 0x2000,
    RESERVED_MASK = 0x9500,
    RTSPECIAL_NAME = 0x0400,
    HAS_FIELD_MARSHAL = 0x1000,
    HAS_DEFAULT = 0x8000,
    HAS_FIELD_RVA = 0x0100,
}